+++ M.A.G.U.S/Bestiary/Animals/GrayDolphin.cs
@@ -28,8 +28,8 @@ public sealed class GrayDolphin : Creature
 
     [DiceThrow(ThrowType._1D6)]
     [DiceThrowModifier(2)]
-    public override byte GetDamage() => (byte)(DiceThrow._1D6() + 2);
+    public override int GetDamage() => DiceThrow._1D6() + 2;
 
     [DiceThrow(ThrowType._2D6)]
-    public override byte GetNumberAppearing() => (byte)DiceThrow._2D6();
+    public override int GetNumberAppearing() => DiceThrow._2D6();
 }
