+++ M.A.G.U.S/Bestiary/Undead/Werewolf.cs
@@ -25,9 +25,9 @@ public sealed class Werewolf : LivingDead
     }
 
     [DiceThrow(ThrowType._1D10)]
-    public override byte GetDamage() => (byte)DiceThrow._1D10();
+    public override int GetDamage() => DiceThrow._1D10();
 
-    public override byte GetNumberAppearing() => 1;
+    public override int GetNumberAppearing() => 1;
 
     public override string[] Sounds => ["werewolf_howl"];
 }
