+++ M.A.G.U.S/Classes/Believer/GodsOfPyarron/KyelPriest.cs
@@ -10,7 +10,7 @@ public class KyelPriest : Priest
 {
     public KyelPriest() : base() { }
 
-    public KyelPriest(byte level) : base(level) { }
+    public KyelPriest(int level) : base(level) { }
 
     public override QualificationList Qualifications
     {
