+++ M.A.G.U.S/Classes/Believer/GodsOfPyarron/VelarPaladin.cs
@@ -17,81 +17,81 @@ public class VelarPaladin : Class, IClass, ILikeMagic
 {
     public VelarPaladin() : base(1) { }
 
-    public VelarPaladin(byte level) : base(level)
+    public VelarPaladin(int level) : base(level)
     {
         GenerateSkills();
     }
 
     [DiceThrow(ThrowType._2D6)]
     [DiceThrowModifier(6)]
-    public override sbyte Strength { get; set; }
+    public override int Strength { get; set; }
 
     [DiceThrow(ThrowType._3D6_2_Times)]
-    public override sbyte Quickness { get; set; }
+    public override int Quickness { get; set; }
 
     [DiceThrow(ThrowType._3D6_2_Times)]
-    public override sbyte Dexterity { get; set; }
+    public override int Dexterity { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
     [DiceThrowModifier(6)]
-    public override sbyte Stamina { get; set; }
+    public override int Stamina { get; set; }
 
     [DiceThrow(ThrowType._1D10)]
     [DiceThrowModifier(8)]
-    public override sbyte Health { get; set; }
+    public override int Health { get; set; }
 
     [DiceThrow(ThrowType._1D10)]
     [DiceThrowModifier(10)]
-    public override sbyte Beauty { get; set; }
+    public override int Beauty { get; set; }
 
     [DiceThrow(ThrowType._1D10)]
     [DiceThrowModifier(8)]
-    public override sbyte Intelligence { get; set; }
+    public override int Intelligence { get; set; }
 
     [DiceThrow(ThrowType._1D10)]
     [DiceThrowModifier(8)]
     [SpecialTraining]
-    public override sbyte Willpower { get; set; }
+    public override int Willpower { get; set; }
 
     [DiceThrow(ThrowType._1D6)]
     [DiceThrowModifier(12)]
     [SpecialTraining]
-    public override sbyte Astral { get; set; }
+    public override int Astral { get; set; }
 
     [DiceThrow(ThrowType._1D6)]
-    public override byte Gold { get; set; }
+    public override int Gold { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
     [DiceThrowModifier(8)]
-    public override sbyte Bravery { get; set; }
+    public override int Bravery { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
     [DiceThrowModifier(8)]
-    public override sbyte Erudition { get; set; }
+    public override int Erudition { get; set; }
 
     [DiceThrow(ThrowType._1D10)]
     [DiceThrowModifier(8)]
-    public override sbyte Detection { get; set; }
+    public override int Detection { get; set; }
 
-    public override byte InitiatingBaseValue => 5;
+    public override int InitiatingBaseValue => 5;
 
-    public override byte AttackingBaseValue => 20;
+    public override int AttackingBaseValue => 20;
 
-    public override byte DefendingBaseValue => 75;
+    public override int DefendingBaseValue => 75;
 
-    public override byte AimingBaseValue => 0;
+    public override int AimingBaseValue => 0;
 
-    public override byte FightValueModifier => 9;
+    public override int FightValueModifier => 9;
 
-    public override byte BaseQualificationPoints => 5;
+    public override int BaseQualificationPoints => 5;
 
-    public override byte QualificationPointsModifier => 5;
+    public override int QualificationPointsModifier => 5;
 
-    public override byte PercentQualificationModifier => 0;
+    public override int PercentQualificationModifier => 0;
 
-    public override byte BaseLifePoints => 8;
+    public override int BaseLifePoints => 8;
 
-    public override byte BasePainTolerancePoints => 7;
+    public override int BasePainTolerancePoints => 7;
 
     public override bool AddFightValueOnFirstLevel => false;
 
@@ -144,7 +144,7 @@ public class VelarPaladin : Class, IClass, ILikeMagic
 
     [DiceThrow(ThrowType._1D6)]
     [DiceThrowModifier(3)]
-    public override byte GetPainToleranceModifier() => (byte)(DiceThrow._1D6() + 3);
+    public override int GetPainToleranceModifier() => DiceThrow._1D6() + 3;
 
     public override string Name => "Paladin of Velar";
 }
