+++ M.A.G.U.S/Classes/NonPlayableCharacters/Peasant.cs
@@ -11,70 +11,70 @@ public class Peasant : Class, IClass
 {
     public Peasant() : base(1) { }
 
-    public Peasant(byte level) : base(level)
+    public Peasant(int level) : base(level)
     {
         GenerateSkills();
     }
 
     [DiceThrow(ThrowType._3D6)]
-    public override sbyte Strength { get; set; }
+    public override int Strength { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
-    public override sbyte Quickness { get; set; }
+    public override int Quickness { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
-    public override sbyte Dexterity { get; set; }
+    public override int Dexterity { get; set; }
 
     [DiceThrow(ThrowType._3D6)]
-    public override sbyte Stamina { get; set; }
+    public override int Stamina { get; set; }
 
     [DiceThrow(ThrowType._3D6)]
-    public override sbyte Health { get; set; }
+    public override int Health { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
-    public override sbyte Beauty { get; set; }
+    public override int Beauty { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
-    public override sbyte Intelligence { get; set; }
+    public override int Intelligence { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
-    public override sbyte Willpower { get; set; }
+    public override int Willpower { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
-    public override sbyte Astral { get; set; }
+    public override int Astral { get; set; }
 
     [DiceThrow(ThrowType._1D2)]
-    public override byte Gold { get; set; }
+    public override int Gold { get; set; }
 
     [DiceThrow(ThrowType._1D6)]
-    public override sbyte Bravery { get; set; }
+    public override int Bravery { get; set; }
 
     [DiceThrow(ThrowType._1D6)]
-    public override sbyte Erudition { get; set; }
+    public override int Erudition { get; set; }
 
     [DiceThrow(ThrowType._2D6)]
     [DiceThrowModifier(6)]
-    public override sbyte Detection { get; set; }
+    public override int Detection { get; set; }
 
-    public override byte InitiatingBaseValue => 1;
+    public override int InitiatingBaseValue => 1;
 
-    public override byte AttackingBaseValue => 5;
+    public override int AttackingBaseValue => 5;
 
-    public override byte DefendingBaseValue => 50;
+    public override int DefendingBaseValue => 50;
 
-    public override byte AimingBaseValue => 0;
+    public override int AimingBaseValue => 0;
 
-    public override byte FightValueModifier => 4;
+    public override int FightValueModifier => 4;
 
-    public override byte BaseQualificationPoints => 0;
+    public override int BaseQualificationPoints => 0;
 
-    public override byte QualificationPointsModifier => 0;
+    public override int QualificationPointsModifier => 0;
 
-    public override byte PercentQualificationModifier => 0;
+    public override int PercentQualificationModifier => 0;
 
-    public override byte BaseLifePoints => 3;
+    public override int BaseLifePoints => 3;
 
-    public override byte BasePainTolerancePoints => 10;
+    public override int BasePainTolerancePoints => 10;
 
     public override bool AddFightValueOnFirstLevel => false;
 
@@ -97,5 +97,5 @@ public class Peasant : Class, IClass
     public override SpecialQualificationList SpecialQualifications => [];
 
     [DiceThrow(ThrowType._1D2)]
-    public override byte GetPainToleranceModifier() => (byte)DiceThrow._1D2();
+    public override int GetPainToleranceModifier() => DiceThrow._1D2();
 }
