+++ M.A.G.U.S/GameSystem/Attributes/DiceThrowModifierAttribute.cs
@@ -1,7 +1,7 @@
 ﻿namespace M.A.G.U.S.GameSystem.Attributes;
 
 [AttributeUsage(AttributeTargets.Property | AttributeTargets.Method)]
-public class DiceThrowModifierAttribute(byte modifier) : Attribute
+public class DiceThrowModifierAttribute(int modifier) : Attribute
 {
-    public byte Modifier { get; } = modifier;
+    public int Modifier { get; } = modifier;
 }
